+++ src/components/base/b-virtual-scroll-new/modules/state/index.ts
@@ -104,7 +104,7 @@ export class ComponentInternalState extends Friend {
 
 		const
 			chunkSize = ctx.getChunkSize(state),
-			dataOffset = this.getDataCursor() + chunkSize;
+			dataOffset = this.getDataOffset() + chunkSize;
 
 		if (<CanUndef<object>>state.data[dataOffset] == null) {
 			state.isLastRender = true;
@@ -125,6 +125,14 @@ export class ComponentInternalState extends Friend {
 	 */
 	setIsInitialRender(value: boolean): void {
 		this.state.isInitialRender = value;
+	}
+
+	/**
+	 * Sets the flag indicating that the process of inserting components into the DOM tree is currently in progress
+	 * @param value
+	 */
+	setIsDomInsertInProgress(value: boolean): void {
+		this.privateState.isDomInsertInProgress = value;
 	}
 
 	/**
@@ -186,8 +194,16 @@ export class ComponentInternalState extends Friend {
 	/**
 	 * Returns the cursor indicating the last index of the last rendered data element
 	 */
-	getDataCursor(): number {
+	getDataOffset(): number {
 		return this.privateState.dataOffset;
+	}
+
+	/**
+	 * Returns the value of the flag indicating whether the process
+	 * of inserting components into the DOM tree is currently in progress
+	 */
+	getIsDomInsertInProgress(): boolean {
+		return this.privateState.isDomInsertInProgress;
 	}
 
 	/**
@@ -196,7 +212,7 @@ export class ComponentInternalState extends Friend {
 	updateDataOffset(): void {
 		const
 			{ctx, state} = this,
-			current = this.getDataCursor(),
+			current = this.getDataOffset(),
 			chunkSize = ctx.getChunkSize(state);
 
 		this.privateState.dataOffset = current + chunkSize;