+++ src/core/component/directives/ref/index.ts
@@ -112,7 +112,7 @@ function updateRef(el: Element | ComponentElement, opts: DirectiveOptions, vnode
 			refVal = getRefVal();
 
 		let
-			ref;
+			ref: unknown;
 
 		if (Object.isArray(refVal)) {
 			if (key != null) {