+++ tests/helpers/index.ts
@@ -11,9 +11,12 @@ import BOM from 'tests/helpers/bom';
 import Utils from 'tests/helpers/utils';
 import Assert from 'tests/helpers/assert';
 import Component from 'tests/helpers/component';
+import ComponentObject from 'tests/helpers/component-object';
 import Scroll from 'tests/helpers/scroll';
 import Router from 'tests/helpers/router';
 import Gestures from 'tests/helpers/gestures';
+import RequestInterceptor from 'tests/helpers/network/interceptor';
+import * as Mock from 'tests/helpers/mock';
 
 export {
 
@@ -22,8 +25,11 @@ export {
 	Utils,
 	Assert,
 	Component,
+	ComponentObject,
 	Scroll,
 	Router,
-	Gestures
+	Gestures,
+	RequestInterceptor,
+	Mock
 
 };