+++ src/core/component/interface/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.62 (2024-02-19)
+
+#### :rocket: New Feature
+
+* Added an app property to get a reference to the application object
+
 ## v4.0.0-beta.32 (2023-10-17)
 
 #### :rocket: New Feature