+++ src/core/data/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v4.0.0-beta.57 (2024-02-13)
+
+#### :house: Internal
+
+* Removed `id` property from `ProviderOptions` interface
+
 ## v4.0.0-beta.43 (2023-11-26)
 
 #### :bug: Bug Fix